Reverted commit e1436c7d (move mtr to InnoDB). The problem is that we have several tests that call add_suppression procedure inside transaction. It has several problems: 1. Now mixing InnoDB and MyIsam emits a warning 2. In general we don't want to replicate insertions done by add_suppresion. So we should set wsrep_on=OFF - INSERT - wsrep_on=ON. Several probles here:   1. It is not allowed to switch wsrep_on within transaction   2. It affects binlog 3. We could allow replication, but what is really desireable to avoid is to not binlog add_suppression to not affect (many!) tests that operate on binlog. But toggling sql_log_bin in active transaction is not allowed as well.
For now, just re-record tests that emit the warning.

+++ mysql-test/include/mtr_warnings.sql
@@ -33,15 +33,11 @@ use mtr;
 -- Create table where testcases can insert patterns to
 -- be suppressed
 --
--- with PXC we want to avoid replication
+-- with PXC we keep this table in MyISAM to avoid replication
 -- of suppression added to one node to other nodes of the cluster.
-
-SET @wsrep_on_saved = @@session.wsrep_on;
-SET @@session.wsrep_on = OFF;
 CREATE TABLE test_suppressions (
   pattern VARCHAR(255) NOT NULL
-);
-SET @@session.wsrep_on = @wsrep_on_saved;
+) engine=MyISAM;
 
 --
 -- Table of full messages (not patterns), suppressed by a test while it
@@ -49,15 +45,12 @@ SET @@session.wsrep_on = @wsrep_on_saved;
 -- Primary key is guaranteed to be unique because the prefix includes the
 -- timestamp, which the server guarantees (!) is unique.
 --
--- with PXC we want to avoid replication
+-- with PXC we keep this table in MyISAM to avoid replication
 -- of suppression added to one node to other nodes of the cluster.
-SET @wsrep_on_saved = @@session.wsrep_on;
-SET @@session.wsrep_on = OFF;
 CREATE TABLE asserted_test_suppressions (
   message TEXT NOT NULL,
   PRIMARY KEY(message(100))
-);
-SET @@session.wsrep_on = @wsrep_on_saved;
+) engine=MyISAM;
 
 --
 -- Table of patterns for messages that should be excluded from global
@@ -70,15 +63,12 @@ SET @@session.wsrep_on = @wsrep_on_saved;
 -- global suppression pattern. Instead, the pattern should match the
 -- error messages for which global suppression should be ignored.
 --
--- with PXC we want to avoid replication
+-- with PXC we keep this table in MyISAM to avoid replication
 -- of suppression added to one node to other nodes of the cluster.
-SET @wsrep_on_saved = @@session.wsrep_on;
-SET @@session.wsrep_on = OFF;
 CREATE TABLE test_ignored_global_suppressions (
   pattern VARCHAR(255) NOT NULL,
   PRIMARY KEY(pattern(255))
-);
-SET @@session.wsrep_on = @wsrep_on_saved;
+) engine=MyISAM;
 
 --
 -- Declare a trigger that makes sure
@@ -479,18 +469,17 @@ INSERT INTO global_suppressions VALUES
  ("Using group replication with Percona XtraDB Cluster is only supported for migration"),
 
  /*
-  KH: TODO: remove below suppressions after testing
    Warnings from Clone plugin
 
    In PXC, the table mtr.test_suppressions is created as MyISAM to avoid
    replication of suppression added to one node to other nodes of the cluster.
-
+ */
  ("Non innodb table: mtr.test_suppressions is not cloned and is empty."),
  ("Clone removing all user data for provisioning: Started"),
  ("Clone removing all user data for provisioning: Finished"),
  ("\\[Warning\\] .*Non innodb table: .* is not cloned and is empty."),
  ("\\[ERROR\\] .*MY-\\d+.*clone_check_recovery_crashpoint.*"),
-*/
+
  /*
    Warnings/errors seen when server is loaded with keyring plugin without
    enabling pxc_encrypt_cluster_traffic.
@@ -515,11 +504,8 @@ BEGIN
   -- Protect the mark on lines that match an 'ignore suppression' pattern.
   --
   SET GLOBAL regexp_time_limit = 0;
-  SET @wsrep_on_saved = @@session.wsrep_on;
-  SET @@session.wsrep_on = OFF;
   UPDATE error_log el, test_ignored_global_suppressions igs
     SET suspicious = 2 WHERE el.line REGEXP igs.pattern;
-  SET @@session.wsrep_on = @wsrep_on_saved;
   --
   -- Remove the mark from lines that are suppressed by global suppressions.
   --
@@ -538,12 +524,9 @@ BEGIN
   -- Remove mark from lines that are suppressed by test specific suppressions
   --
   SET GLOBAL regexp_time_limit = 0;
-  SET @wsrep_on_saved = @@session.wsrep_on;
-  SET @@session.wsrep_on = OFF;
   UPDATE error_log el, test_suppressions ts
     SET suspicious=0
       WHERE el.suspicious=1 AND el.line REGEXP ts.pattern;
-  SET @@session.wsrep_on = @wsrep_on_saved;
   SET GLOBAL regexp_time_limit = DEFAULT;
 END$$
 
@@ -556,12 +539,9 @@ BEGIN
   -- context.
   --
   SET GLOBAL regexp_time_limit = 0;
-  SET @wsrep_on_saved = @@session.wsrep_on;
-  SET @@session.wsrep_on = OFF;
   UPDATE error_log el, asserted_test_suppressions ats
     SET suspicious=0
       WHERE el.suspicious=1 AND el.line = ats.message;
-  SET @@session.wsrep_on = @wsrep_on_saved;
   SET GLOBAL regexp_time_limit = DEFAULT;
 END$$
 
@@ -596,15 +576,20 @@ BEGIN
   END IF;
 
   -- Cleanup for next test
-  -- The TRUNCATE should not be replicated under Galera
-  -- as it causes the custom suppressions on the other
-  -- nodes to be deleted as well
-  SET @wsrep_on_saved = @@session.wsrep_on;
-  SET @@session.wsrep_on = OFF;
-  TRUNCATE test_suppressions;
-  TRUNCATE test_ignored_global_suppressions;
-  TRUNCATE asserted_test_suppressions;
-  SET @@session.wsrep_on = @wsrep_on_saved;
+  IF @@wsrep_on = 1 THEN
+    -- The TRUNCATE should not be replicated under Galera
+    -- as it causes the custom suppressions on the other
+    -- nodes to be deleted as well
+    SET wsrep_on = 0;
+    TRUNCATE test_suppressions;
+    TRUNCATE test_ignored_global_suppressions;
+    TRUNCATE asserted_test_suppressions;
+    SET wsrep_on = 1;
+  ELSE
+    TRUNCATE test_suppressions;
+    TRUNCATE test_ignored_global_suppressions;
+    TRUNCATE asserted_test_suppressions;
+  END IF;
 
   DROP TABLE error_log;
 
@@ -618,14 +603,10 @@ END$$
 CREATE DEFINER=root@localhost
 PROCEDURE add_suppression(pattern VARCHAR(255))
 BEGIN
-  SET @wsrep_on_saved = @@session.wsrep_on;
-  SET @@session.wsrep_on = OFF;
   INSERT INTO test_suppressions (pattern) VALUES (pattern);
-  SET @@session.wsrep_on = @wsrep_on_saved;
 END
 */$$
 
 DELIMITER ;
-
